src/deploy_requirements.txt
google-cloud-aiplatform
google-cloud-storage
torch
tiktoken